+++ subsys/bluetooth/host/hfp_hf.c
@@ -21,7 +21,7 @@
 
 #include "hci_core.h"
 #include "conn_internal.h"
-#include "l2cap_internal.h"
+#include "l2cap_br_internal.h"
 #include "rfcomm_internal.h"
 #include "at.h"
 #include "hfp_internal.h"
