+++ subsys/bluetooth/host/l2cap_internal.h
@@ -10,6 +10,7 @@
 
 #include <zephyr/bluetooth/l2cap.h>
 #include <zephyr/sys/iterable_sections.h>
+#include "l2cap_br_interface.h"
 
 enum l2cap_conn_list_action {
 	BT_L2CAP_CHAN_LOOKUP,
@@ -50,64 +51,6 @@ struct bt_l2cap_cmd_reject_cid_data {
 	uint16_t dcid;
 } __packed;
 
-#define BT_L2CAP_CONN_REQ               0x02
-struct bt_l2cap_conn_req {
-	uint16_t psm;
-	uint16_t scid;
-} __packed;
-
-/* command statuses in response */
-#define BT_L2CAP_CS_NO_INFO             0x0000
-#define BT_L2CAP_CS_AUTHEN_PEND         0x0001
-
-/* valid results in conn response on BR/EDR */
-#define BT_L2CAP_BR_SUCCESS             0x0000
-#define BT_L2CAP_BR_PENDING             0x0001
-#define BT_L2CAP_BR_ERR_PSM_NOT_SUPP    0x0002
-#define BT_L2CAP_BR_ERR_SEC_BLOCK       0x0003
-#define BT_L2CAP_BR_ERR_NO_RESOURCES    0x0004
-#define BT_L2CAP_BR_ERR_INVALID_SCID    0x0006
-#define BT_L2CAP_BR_ERR_SCID_IN_USE     0x0007
-
-#define BT_L2CAP_CONN_RSP               0x03
-struct bt_l2cap_conn_rsp {
-	uint16_t dcid;
-	uint16_t scid;
-	uint16_t result;
-	uint16_t status;
-} __packed;
-
-#define BT_L2CAP_CONF_SUCCESS           0x0000
-#define BT_L2CAP_CONF_UNACCEPT          0x0001
-#define BT_L2CAP_CONF_REJECT            0x0002
-
-#define BT_L2CAP_CONF_REQ               0x04
-struct bt_l2cap_conf_req {
-	uint16_t dcid;
-	uint16_t flags;
-	uint8_t  data[0];
-} __packed;
-
-#define BT_L2CAP_CONF_RSP               0x05
-struct bt_l2cap_conf_rsp {
-	uint16_t scid;
-	uint16_t flags;
-	uint16_t result;
-	uint8_t  data[0];
-} __packed;
-
-/* Option type used by MTU config request data */
-#define BT_L2CAP_CONF_OPT_MTU           0x01
-/* Options bits selecting most significant bit (hint) in type field */
-#define BT_L2CAP_CONF_HINT              0x80
-#define BT_L2CAP_CONF_MASK              0x7f
-
-struct bt_l2cap_conf_opt {
-	uint8_t type;
-	uint8_t len;
-	uint8_t data[0];
-} __packed;
-
 #define BT_L2CAP_DISCONN_REQ            0x06
 struct bt_l2cap_disconn_req {
 	uint16_t dcid;
@@ -118,25 +61,6 @@ struct bt_l2cap_disconn_req {
 struct bt_l2cap_disconn_rsp {
 	uint16_t dcid;
 	uint16_t scid;
-} __packed;
-
-#define BT_L2CAP_INFO_FEAT_MASK         0x0002
-#define BT_L2CAP_INFO_FIXED_CHAN        0x0003
-
-#define BT_L2CAP_INFO_REQ               0x0a
-struct bt_l2cap_info_req {
-	uint16_t type;
-} __packed;
-
-/* info result */
-#define BT_L2CAP_INFO_SUCCESS           0x0000
-#define BT_L2CAP_INFO_NOTSUPP           0x0001
-
-#define BT_L2CAP_INFO_RSP               0x0b
-struct bt_l2cap_info_rsp {
-	uint16_t type;
-	uint16_t result;
-	uint8_t  data[0];
 } __packed;
 
 #define BT_L2CAP_CONN_PARAM_REQ         0x12
@@ -243,20 +167,6 @@ struct bt_l2cap_fixed_chan {
 				.destroy = _destroy,                    \
 			}
 
-/* Need a name different than bt_l2cap_fixed_chan for a different section */
-struct bt_l2cap_br_fixed_chan {
-	uint16_t		cid;
-	int (*accept)(struct bt_conn *conn, struct bt_l2cap_chan **chan);
-};
-
-#define BT_L2CAP_BR_CHANNEL_DEFINE(_name, _cid, _accept)		\
-	const STRUCT_SECTION_ITERABLE(bt_l2cap_br_fixed_chan, _name) = { \
-				.cid = _cid,			\
-				.accept = _accept,		\
-			}
-
-#define BR_CHAN(_ch) CONTAINER_OF(_ch, struct bt_l2cap_br_chan, chan)
-
 /* Notify L2CAP channels of a new connection */
 void bt_l2cap_connected(struct bt_conn *conn);
 
@@ -300,9 +210,6 @@ struct net_buf *bt_l2cap_create_pdu_timeout(struct net_buf_pool *pool,
 #define bt_l2cap_create_pdu(_pool, _reserve) \
 	bt_l2cap_create_pdu_timeout(_pool, _reserve, K_FOREVER)
 
-/* Prepare a L2CAP Response PDU to be sent over a connection */
-struct net_buf *bt_l2cap_create_rsp(struct net_buf *buf, size_t reserve);
-
 /* Send L2CAP PDU over a connection
  *
  * Buffer ownership is transferred to stack in case of success.
@@ -333,40 +240,6 @@ struct bt_l2cap_chan *bt_l2cap_le_lookup_tx_cid(struct bt_conn *conn,
 /* Lookup channel by Receiver CID */
 struct bt_l2cap_chan *bt_l2cap_le_lookup_rx_cid(struct bt_conn *conn,
 						uint16_t cid);
-
-/* Initialize BR/EDR L2CAP signal layer */
-void bt_l2cap_br_init(void);
-
-/* Register fixed channel */
-void bt_l2cap_br_fixed_chan_register(struct bt_l2cap_fixed_chan *chan);
-
-/* Notify BR/EDR L2CAP channels about established new ACL connection */
-void bt_l2cap_br_connected(struct bt_conn *conn);
-
-/* Lookup BR/EDR L2CAP channel by Receiver CID */
-struct bt_l2cap_chan *bt_l2cap_br_lookup_rx_cid(struct bt_conn *conn,
-						uint16_t cid);
-
-/* Disconnects dynamic channel */
-int bt_l2cap_br_chan_disconnect(struct bt_l2cap_chan *chan);
-
-/* Make connection to peer psm server */
-int bt_l2cap_br_chan_connect(struct bt_conn *conn, struct bt_l2cap_chan *chan,
-			     uint16_t psm);
-
-/* Send packet data to connected peer */
-int bt_l2cap_br_chan_send(struct bt_l2cap_chan *chan, struct net_buf *buf);
-int bt_l2cap_br_chan_send_cb(struct bt_l2cap_chan *chan, struct net_buf *buf, bt_conn_tx_cb_t cb,
-			     void *user_data);
-
-/*
- * Handle security level changed on link passing HCI status of performed
- * security procedure.
- */
-void l2cap_br_encrypt_change(struct bt_conn *conn, uint8_t hci_status);
-
-/* Handle received data */
-void bt_l2cap_br_recv(struct bt_conn *conn, struct net_buf *buf);
 
 struct bt_l2cap_ecred_cb {
 	void (*ecred_conn_rsp)(struct bt_conn *conn, uint16_t result, uint8_t attempted,
